[Refactoring] Create repository.thread_list package.
