+++ app/src/main/java/com/u1fukui/bbs/repository/thread_list/HistoryThreadListRepository.kt
@@ -1,4 +1,4 @@
-package com.u1fukui.bbs.repository
+package com.u1fukui.bbs.repository.thread_list
 
 
 import com.u1fukui.bbs.api.ThreadListApi
@@ -9,11 +9,9 @@ import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.schedulers.Schedulers
 import java.util.*
-import javax.inject.Inject
 
-class CategoryThreadListRepository @Inject constructor(
-        private val threadListApi: ThreadListApi,
-        private val categoryId: Int
+class HistoryThreadListRepository constructor(
+        private val threadListApi: ThreadListApi
 ) : ThreadListRepository {
 
     override fun fetchThreadList(lastId: Long): Single<ThreadListResponse> =
@@ -21,7 +19,7 @@ class CategoryThreadListRepository @Inject constructor(
                     .subscribeOn(Schedulers.io())
                     .observeOn(AndroidSchedulers.mainThread())
                     .map {
-                        ThreadListResponse(0,
+                        ThreadListResponse(200,
                                 createDebugDataList(lastId),
                                 lastId >= 100)
                     }
@@ -29,9 +27,9 @@ class CategoryThreadListRepository @Inject constructor(
     private fun createDebugDataList(lastId: Long): List<BbsThread> =
             ((lastId + 1)..(lastId + 20)).map {
                 BbsThread(
-                        it.toLong(),
-                        "カテゴリスレッド$it",
-                        User(it.toLong(), "作者$it"),
+                        it,
+                        "履歴スレッド$it",
+                        User(it, "作者$it"),
                         0,
                         Date(),
                         Date()
