+++ app/src/main/java/com/u1fukui/bbs/ui/creation/thread/SelectCategoryViewModel.kt
@@ -6,7 +6,7 @@ import android.databinding.ObservableList
 import com.u1fukui.bbs.customview.ErrorView
 import com.u1fukui.bbs.helper.LoadingManager
 import com.u1fukui.bbs.model.Category
-import com.u1fukui.bbs.repository.CategoryListRepository
+import com.u1fukui.bbs.repository.thread_list.CategoryListRepository
 import com.u1fukui.bbs.ui.ViewModel
 import io.reactivex.SingleObserver
 import io.reactivex.android.schedulers.AndroidSchedulers
