+++ app/src/main/java/com/u1fukui/bbs/ui/main/mypage/FavoriteThreadListFragment.kt
@@ -1,7 +1,7 @@
 package com.u1fukui.bbs.ui.main.mypage
 
 import com.u1fukui.bbs.api.ThreadListApi
-import com.u1fukui.bbs.repository.FavoriteThreadListRepository
+import com.u1fukui.bbs.repository.thread_list.FavoriteThreadListRepository
 import com.u1fukui.bbs.ui.main.BaseThreadListFragment
 import javax.inject.Inject
